Extract worker virtual DOM publisher
